combine metadata, schema_view, and schema
